# server/docs/swagger.yaml
definitions:
...
  rest.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  rest.EventData:
    properties:
      data:
        properties:
          email_addresses:
            items:
              properties:
                email_address:
                  type: string
              type: object
            type: array
          id:
            type: string
        type: object
    type: object
...
  rest.PurchaseRequest:
    properties:
      product_ids:
        items:
          type: string
        type: array
      user_id:
        type: string
    type: object
  rest.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/users/create:
    post:
      consumes:
      - application/json
      description: Create a new user with user ID and email address
      parameters:
      - description: Event data
        in: body
        name: eventData
        required: true
        schema:
          $ref: '#/definitions/rest.EventData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Create a new user
      tags:
      - user
...
  /api/users/purchase:
    post:
      consumes:
      - application/json
      description: Create a purchase record for a user
      parameters:
      - description: Purchase Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.PurchaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Create a purchase
      tags:
      - user
swagger: "2.0"